This initial commit includes a basic implementation of the Guess Who game.
The game currently supports two modes: 'bot_vs_bot' and 'vs_human'. In 'bot_vs_bot' mode, the game autonomously selects characters, asks questions, and makes guesses. In 'vs_human' mode, the user manually enters their characters and responds to the bot's questions.

The game uses a basic strategy to choose the best question by striving to maintain a balance in the distribution of attributes among the remaining possible characters.

As an extension to the traditional Guess Who game, we've begun to incorporate support for an advanced mode where each player has two cards instead of one. This introduces a new layer of complexity as the bot now has to strategize over two possible characters simultaneously.

While the current structure of the game has been adapted to account for this multi-card setup, further development and testing are required to fine-tune the game dynamics and the bot's strategy. We are actively working on improving the advanced version of the game to ensure a challenging and fun experience for the players.

Any contributions to improve the bot's guessing strategy, the game's interactive elements or the overall game logic are  welcome.